requirements.txt
opencv-python
imread-from-url
onnxruntime-gpu
cap-from-youtube
rerun-sdk